created chakra ui theme file
